remove unecessary semicolon

+++ challenge-1/challenge.ts
@@ -4,9 +4,7 @@
  * This is the core function you'll need to edit, though you're encouraged to make helper
  * functions!
  */
-import https from 'https';
 import fs from 'fs';
-import {IncomingMessage} from "http";
 import zlib from 'zlib';
 import tar from 'tar';
 import knex, {Knex} from 'knex';
@@ -15,34 +13,11 @@ import {DUMP_DOWNLOAD_URL, SQLITE_DB_PATH} from "./resources";
 import path from "path";
 import _ from "lodash";
 import axios from 'axios';
-import { Writable } from 'stream';
+import assert from "assert";
+import Dict = NodeJS.Dict;
 
 
 type InferredType = 'int' | 'bigint' | 'float' | 'date' | 'string';
-
-// /**
-//  * Downloads a file from a given URL and saves it to a specified destination.
-//  * @param {string} url - The URL of the file to download.
-//  * @param {string} dest - The destination where the file should be saved.
-//  * @returns {Promise<void>} A promise that resolves when the download is complete.
-//  */
-// const downloadFile = (url: string, dest: string): Promise<void> => {
-//   return new Promise((resolve, reject) => {
-//     const file = fs.createWriteStream(dest);
-//     https.get(url, (response: IncomingMessage) => {
-//       response.pipe(file);
-//       file.on('finish', () => {
-//         file.close();
-//         resolve();
-//       });
-//     }).on('error', (err: Error) => {
-//       fs.unlink(dest, unlinkErr => {
-//         if (unlinkErr) console.error(`Failed to unlink file: ${unlinkErr}`);
-//       });
-//       reject(err);
-//     });
-//   });
-// };
 
 /**
  * Downloads a file from a given URL and saves it to a specified destination.
@@ -133,7 +108,7 @@ const inferColumnTypes = async (filePath: string, sampleSize: number = 10): Prom
   const sampleRows: any[] = [];
   const inferredTypes: Record<string, InferredType> = {};
 
-  return new Promise((resolve, reject) => {
+  return new Promise((resolve, reject): void => {
     fs.createReadStream(filePath)
         .pipe(csv.parse({ headers: true }))
         .on('data', (row) => {
@@ -141,13 +116,13 @@ const inferColumnTypes = async (filePath: string, sampleSize: number = 10): Prom
             sampleRows.push(row);
           }
         })
-        .on('end', () => {
+        .on('end', (): void => {
           // Infer types based on sample rows
           const headers = Object.keys(sampleRows[0]);
-          headers.forEach((header) => {
-            const sampleValues = sampleRows.map(row => row[header]);
-            const types = new Set(sampleValues.map(inferType));
-            inferredTypes[header] = Array.from(types)[0];
+          _.forEach(headers, (header) => {
+            const sampleValues = _.map(sampleRows, row => row[header]);
+            const types: Set<InferredType> = new Set(sampleValues.map(inferType));
+            inferredTypes[header] =  Array.from(types)[0];
           });
           resolve(inferredTypes);
         })
@@ -170,9 +145,9 @@ const setupDatabase = async (inferredTypes: Record<string, InferredType>, tableN
     useNullAsDefault: true,
   });
 
-  await db.schema.createTable(tableName, (table: Knex.CreateTableBuilder) => {
+  await db.schema.createTable(tableName, (table: Knex.CreateTableBuilder): void => {
     table.increments('id').primary();
-    _.forEach(Object.entries(inferredTypes), ([header, type]) => {
+    _.forEach(Object.entries(inferredTypes), ([header, type]): void => {
       switch (type) {
         case 'int':
           table.integer(header);
@@ -209,83 +184,125 @@ const processCSV = async (db: Knex, filePath: string, tableName: string): Promis
   const batchSize: number = 200;
 
   try {
-    // Start a transaction
-    await db.transaction(async (trx) => {
-      return new Promise<void>(async (resolve, reject) => {
-        const stream = fs.createReadStream(filePath)
+    await db.transaction(async (trx): Promise<void> => {
+      return new Promise<void>(async (resolve, reject): Promise<void> => {
+        fs.createReadStream(filePath)
             .pipe(csv.parse({ headers: true }))
-            .on('data', async (row) => {
+            .on('data', async (row): Promise<void> => {
               rows.push(row);
-              if (rows.length >= batchSize) {
-                //stream.pause();  // Pause the stream
-                await trx.batchInsert(tableName, rows.splice(0, batchSize), batchSize).catch(reject);
-                //stream.resume();  // Resume the stream
-              }
+              if (rows.length >= batchSize) await trx.batchInsert(tableName, rows.splice(0, batchSize), batchSize).catch(reject);
             })
-            .on('end', async () => {
-              if (rows.length) {
-                await trx.batchInsert(tableName, rows, batchSize).catch(reject);
-              }
+            .on('end', async (): Promise<void> => {
+              if (rows.length) await trx.batchInsert(tableName, rows, batchSize).catch(reject);
               resolve();
             })
             .on('error', reject);
       });
     });
-
   } catch (err) {
     console.error(`An error occurred while processing the CSV: ${err}`);
   }
 };
 
+/**
+ * Count the number of rows in a SQLite table.
+ * @param {Knex} db - The Knex database connection.
+ * @param {string} tableName - The name of the SQLite table.
+ * @returns {Promise<number>} A promise that resolves with the number of rows in the table.
+ */
+const countRowsInTable = async (db: Knex, tableName: string): Promise<number> => {
+  const result: Dict<string | number>[] = await db(tableName).count('* as count');
+  return Number(result[0].count);
+};
 
 /**
- * The main function. Downloads, decompresses, and processes a data dump into SQLite tables.
- * @returns {Promise<void>} A promise that resolves when all operations are complete.
+ * Count the number of rows in a CSV file.
+ * @param {string} filePath - The path to the CSV file.
+ * @returns {Promise<number>} A promise that resolves with the number of rows in the CSV file.
  */
+const countRowsInCSV = (filePath: string): Promise<number> => {
+  let rowCount: number = 0;
+  return new Promise((resolve, reject): void  => {
+    fs.createReadStream(filePath)
+        .pipe(csv.parse({ headers: true }))
+        .on('data', (): void => {
+          rowCount++;
+        })
+        .on('end', (): void  => {
+          resolve(rowCount);
+        })
+        .on('error', reject);
+  });
+};
+
+/**
+ * Checks if a table exists in the SQLite database.
+ * @param {Knex} db - The Knex database connection.
+ * @param {string} tableName - The name of the table to check.
+ * @returns {Promise<boolean>} A promise that resolves with a boolean indicating if the table exists.
+ */
+const checkTableExistence = async (db: Knex, tableName: string): Promise<boolean> => {
+  return await db.schema.hasTable(tableName);
+};
+
+/**
+ * Checks if the directory at the given filepath exists.
+ * @param {string} filePath - The filepath
+ * @returns {void} - Creates the directory if it does not exist and returns nothing.
+ */
+const ensureDirectoryExists = (filePath: string): void => {
+  const directory = path.dirname(filePath);
+  if (!fs.existsSync(directory)) {
+    fs.mkdirSync(directory, { recursive: true });
+  }
+};
+
+/**
+ * Checks if the tables in the sqlite db exists.
+ * @param {any} db - The filepath
+ * @returns {void} - Creates the directory if it does not exist and returns nothing.
+ */const checkAndVerifyTables = async (db: any): Promise<boolean> => {
+  const [doesCustomersTableExist, doesOrganizationsTableExist] = await Promise.all([
+    checkTableExistence(db, 'customers'),
+    checkTableExistence(db, 'organizations'),
+  ]);
+
+  if (doesCustomersTableExist && doesOrganizationsTableExist) {
+    // Verify row counts
+    const [dbCustomerCount, csvCustomerCount, dbOrganizationCount, csvOrganizationCount] = await Promise.all([
+      countRowsInTable(db, 'customers'),
+      countRowsInCSV('./tmp/extracted/dump/customers.csv'),
+      countRowsInTable(db, 'organizations'),
+      countRowsInCSV('./tmp/extracted/dump/organizations.csv'),
+    ]);
+
+    assert(dbCustomerCount !== csvCustomerCount, "customer count in db != customer count in csv");
+    assert(dbOrganizationCount !== csvOrganizationCount, "organization count in db != organization count in csv");
+
+    console.log("Required tables 'customers' and 'organizations' already exist and have the required amount of rows.");
+    await db.destroy();
+    return true;
+  }
+  return false;
+};
+
+// Main function
 export async function processDataDump(): Promise<void> {
   try {
-
-    // Helper function to ensure directory exists
-    const ensureDirectoryExists = (filePath: string) => {
-      const directory = path.dirname(filePath);
-      if (!fs.existsSync(directory)) {
-        fs.mkdirSync(directory, { recursive: true });
-      }
-    };
-
-    // Ensure the directory for SQLite exists
     console.log('Ensuring directory exists for SQLite database...');
     ensureDirectoryExists(SQLITE_DB_PATH);
     console.log('Directory check complete.');
 
-    // Replace these with actual file paths and URLs
     const tarGzUrl = DUMP_DOWNLOAD_URL;
     const tarGzDest = '/tmp/dump.tar.gz';
     const extractDest = './tmp/extracted';
 
-    // Step 1: Download the file
     await downloadFile(tarGzUrl, tarGzDest);
-
-    // Step 2: Decompress and extract the file
     await decompressAndExtract(tarGzDest, extractDest);
 
-    // // Step 3: Infer column types
-    // const customerInferredTypes = await inferColumnTypes('./tmp/extracted/dump/customers.csv');
-    // const organizationInferredTypes = await inferColumnTypes('./tmp/extracted/dump/organizations.csv');
-    //
-    // // Step 4: Setup database
-    // await setupDatabase(customerInferredTypes, 'customers');
-    // await setupDatabase(organizationInferredTypes, 'organizations');
-
-    // Parallelize type inference and database setup
     const [customerInferredTypes, organizationInferredTypes] = await Promise.all([
       inferColumnTypes('./tmp/extracted/dump/customers.csv'),
       inferColumnTypes('./tmp/extracted/dump/organizations.csv')
-    ]);
-
-    await Promise.all([
-      setupDatabase(customerInferredTypes, 'customers'),
-      setupDatabase(organizationInferredTypes, 'organizations')
     ]);
 
     const db = knex({
@@ -296,13 +313,20 @@ export async function processDataDump(): Promise<void> {
       useNullAsDefault: true,
     });
 
-    // // Step 5: Process CSV (Assuming you have a function for it)
+    if (await checkAndVerifyTables(db)) return;
+
+    await Promise.all([
+      setupDatabase(customerInferredTypes, 'customers'),
+      setupDatabase(organizationInferredTypes, 'organizations')
+    ]);
+
     await processCSV(db,'./tmp/extracted/dump/customers.csv', 'customers');
     await processCSV(db, './tmp/extracted/dump/organizations.csv', 'organizations');
 
-    await db.destroy()
+    await db.destroy();
 
   } catch (err) {
     throw Error(`An error occurred: ${err}`);
-  };
-};
+  }
+}
+
